refactor(Auth): Added an API to the event of every Auth api hub events (#473)
Auth category emits two types of hub events - 1) Whenever an api completes 1) For specific Auth events like user signedIn, signedOut or session expired. To distinguish between the two, this PR adds API to the end of hub event names that are fired when an api completes it execution. For example:
When Amplify.Auth.signIn(...) completes, it sends out an event HubPayload.EventName.Auth.signInAPI. On the otherhand when a user is successfully signed in to the Auth category, an event of name HubPayload.EventName.Auth.signedIn will be fired.